+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.142.0] - 2020-03-05
+
+- added `DcaUtil::setFieldsToReadOnly()`
+- added `DcaUtil::getTranslatedModuleNameByTable()`
+
 ## [2.141.1] - 2020-03-03
 
 - added option `restrictFields` to `FormUtil::getModelDataAsNotificationTokens()`